run iterations from sirun subprocesses
`ru_maxrss` is not cumulative but a maximum among all subprocesses. This
makes it relatively useless as soon as there's more than one subprocess.
The workaround is to create a new subprocess for each iteration, each
having its own single subprocess (i.e. the tested program itself). This
way `ru_maxrss` can refer only to the one single subprocess.

In order to transfer the metric data to the parent sirun process, the
existing statsd port is used.