fix startup and teardown
Startup and teardown commands were not _exactly_ being run directly
before and directly after running a benchmark iteration. This has been fixed by
refactoring the code so that the scripts are run *exactly* before and
after the iteration.

This required a huge refactor of main.rs, but hopefully it's now more
readable (and correct).